Allow '~' at beginning of filename by default in safename
Nikolaus Rath has been field testing the 'safename'
security module.  He found that in general the safename
module causes no problems with normal use (for the tests he ran),
except in one case.  He found that Dropbox didn't work with
the original version of safename, because Dropbox creates
file names beginning with '~'.  He reported this problem
on 2016-06-22.

To reduce problems for new users, this commit changes the
safename default to allow '~' at the beginning of a filename.
Note that this can be configured.

Tested-by: Nikolaus Rath <Nikolaus@rath.org>
Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>